requirements.txt
# =============================================================================
# DEPRECATION WARNING:
#
# The file `requirements.txt` does not influence the package dependencies and
# will not be automatically created in the next version of PyScaffold (v4.x).
#
# Please have look at the docs for better alternatives
# (`Dependency Management` section).
# =============================================================================
#
# Add your pinned requirements so that they can be easily installed with:
# pip install -r requirements.txt
# Remember to also add them in setup.cfg but unpinned.
# Example:
# numpy==1.13.3
# scipy==1.0
#
flask==1.1.1
flask_cors==3.0.8
flask_graphql==2.0.0
graphene==2.1.8
sqlalchemy==1.3.11
bcrypt==3.1.7
flask-jwt-extended==3.24.1
password-strength==0.0.3.post2

graphene-sqlalchemy==2.2.2